+++ persistent/list.py
@@ -34,7 +34,10 @@ class PersistentList(UserList, persistent.Persistent):
        Using the `clear` method, or writing ``del inst[:]`` now only
        results in marking the instance as changed if it actually removed
        items.
+    .. versionchanged:: 4.5.2
+       The `copy` method is available on Python 2.
     """
+    __super_getitem = UserList.__getitem__
     __super_setitem = UserList.__setitem__
     __super_delitem = UserList.__delitem__
     __super_iadd = UserList.__iadd__
@@ -52,6 +55,29 @@ class PersistentList(UserList, persistent.Persistent):
         else lambda inst: inst.__delitem__(slice(None, None, None))
     )
 
+    if (2, 7, 0) < sys.version_info[:3] < (3, 7, 4):
+        # Prior to 3.7.4, Python 3 failed to properly
+        # return an instance of the same class.
+        # See https://bugs.python.org/issue27639
+        # and https://github.com/zopefoundation/persistent/issues/112.
+        # We only define the special method on the necessary versions to avoid
+        # any speed penalty.
+        def __getitem__(self, item):
+            result = self.__super_getitem(item)
+            if isinstance(item, slice):
+                result = self.__class__(result)
+            return result
+
+    if sys.version_info[:3] < (3, 7, 4):
+        # Likewise for __copy__.
+        # See https://github.com/python/cpython/commit/3645d29a1dc2102fdb0f5f0c0129ff2295bcd768
+        def __copy__(self):
+            inst = self.__class__.__new__(self.__class__)
+            inst.__dict__.update(self.__dict__)
+            # Create a copy and avoid triggering descriptors
+            inst.__dict__["data"] = self.__dict__["data"][:]
+            return inst
+
     def __setitem__(self, i, item):
         self.__super_setitem(i, item)
         self._p_changed = 1
@@ -67,6 +93,9 @@ class PersistentList(UserList, persistent.Persistent):
     if PYTHON2:  # pragma: no cover
         __super_setslice = UserList.__setslice__
         __super_delslice = UserList.__delslice__
+
+        def copy(self):
+            return self.__class__(self)
 
         def __setslice__(self, i, j, other):
             self.__super_setslice(i, j, other)